Begin transitioning build tool and tests to sbt/scala.
This commit adds a build.sbt file that lets us use sbt for testing.  Eventually
this will transition to using it for the complete build, but it doesn't yet
handle setting resources and parameters correctly, so pom.xml hasn't been
deleted yet.

As a testcase for porting tests, this moves BisquareLinearFitTest.java over to
using scalatest.

Test plan: `sbt test`